[MIOpen] Ported bn_spatial_test from CTest to GTest (#2735)
## Motivation

This is a part of ongoing porting tests from CTest to GTest activity

## Technical Details

The porting was a bit difficult due to the unnecessary complicated
`test_driver` class and its checks. Basically, one has to provide their
own comparison of results instead of generic `test_drive::verify`. Also,
it's really obscure how the values were generated, but I managed to
figure out that basically `get_bn_spatial_input_tensor` has 2 variants:
one for `MIOPEN_TEST_ALL` and one for shorter tests. So I replicated
those using 'Smoke' and 'Full' name conventions in gtest.

## Test Plan

Testing done by simply running test locally, although it won't represent
AMD's Jenkins. So, this PR is actually a part of the testing

## Test Result

See actions for this PR

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.
